pinecone_requirements.txt
langchain==0.0.274
gpt4all==1.0.8
pinecone-client==2.2.2
llama-cpp-python==0.1.81
urllib3==2.0.4
PyMuPDF==1.23.5
python-dotenv==1.0.0
unstructured==0.10.8
extract-msg==0.45.0
tabulate==0.9.0
pandoc==2.3
pypandoc==1.11
tqdm==4.66.1
sentence-transformers==2.2.2
huggingface-hub<0.17.0
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
pinecone-client 
regex 
langchain 
tqdm 
sentence-transformers